rename source folder into src
