+++ setup.cfg
@@ -24,7 +24,7 @@ install_requires =
   overrides==6.1.0
 packages = find:
 package_dir = 
-  mfranceschi_minesweeper = mfranceschi_minesweeper
+  src = mfranceschi_minesweeper
 python_requires = >= 3.7
 
 [options.entry_points]
