+++ src/view/cell_view.py
