+++ tests/__init__.py
@@ -1,3 +1,3 @@
 import sys
 import os
-sys.path.append(os.path.dirname(os.path.realpath(__file__)) + "/../mfranceschi_minesweeper")
+sys.path.append(os.path.dirname(os.path.realpath(__file__)) + "/../src")
